Simple tutorial (without fragments, ViewPagers and vertical layout)
