+++ res/layout/activity_tutorial.xml
@@ -3,22 +3,62 @@
     android:id="@+id/containerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#FFF"
+    android:background="#555"
     android:clipToPadding="false"
     android:fitsSystemWindows="true"
     android:orientation="vertical">
 
-    <android.support.v4.view.ViewPager
-        android:id="@+id/viewPager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+    <LinearLayout
+		android:id="@+id/container"
+		android:layout_width="match_parent"
+		android:layout_height="fill_parent"
+		android:orientation="vertical"
+		android:paddingBottom="60dp"
+		android:weightSum="100">
+
+
+		<ImageView
+			android:id="@+id/image"
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content"
+			android:layout_gravity="center_horizontal"
+			android:layout_marginBottom="16dp"
+			android:layout_weight="20"
+			android:layout_marginTop="16dp"
+			android:adjustViewBounds="true" />
+
+		<TextView
+			android:id="@+id/title"
+			style="@android:style/TextAppearance.DeviceDefault.Large"
+			android:layout_width="match_parent"
+			android:layout_height="match_parent"
+			android:gravity="center_horizontal|bottom"
+			android:paddingEnd="16dp"
+			android:paddingStart="16dp"
+			android:text="Lorem ipsum dolor sit amet"
+			android:textColor="#FFF"
+			android:layout_weight="80" />
+
+		<TextView
+			android:id="@+id/content"
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content"
+			android:layout_marginTop="16dp"
+			android:layout_marginBottom="32dp"
+			android:gravity="center_horizontal"
+			android:paddingEnd="32dp"
+			android:paddingStart="32dp"
+			android:text="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec tempus elementum sem congue laoreet. Sed pulvinar at purus quis aliquet. Quisque feugiat sed leo nec aliquet. In hac habitasse platea dictumst. Vestibulum enim sapien, gravida et tempus vel, feugiat in turpis. Fusce interdum elit et libero rhoncus dapibus"
+			android:textColor="#FFF"
+			android:textSize="14sp"/>
+	</LinearLayout>
 
     <RelativeLayout
         android:id="@+id/buttonContainer"
         android:layout_width="match_parent"
         android:layout_height="60dp"
         android:layout_gravity="bottom"
-        android:background="@android:color/white"
+        android:background="#444"
 		android:clipToPadding="false">
 		<!-- android:elevation="24dp" -->
